1.1.0 beta

+++ src/main/resources/assets/thaumicconcilium/shaders/atsral_frame.frag
@@ -5,11 +5,11 @@
 uniform sampler2D rend_tex;
 uniform int time;
 
-uniform int progress;
-uniform int index;
-uniform int next;
-uniform int[6] indexes;
-uniform int[6] nexts;
+uniform float progress;
+uniform float index;
+uniform float next;
+uniform float[6] indexes;
+uniform float[6] nexts;
 
 vec4 colors[6] = vec4[](
     vec4(1.,0.988,1.,1.),
@@ -29,27 +29,27 @@ void main() {
 
 
     if(color.b == 1.0){
-        color = mix(colors[indexes[0]], colors[nexts[0]], progress / speed) / div;
+        color = mix(colors[int(indexes[0])], colors[int(nexts[0])], progress / speed) / div;
     }
     else if(color.b == 254.0 / 255.0){
-        color = mix(colors[indexes[1]], colors[nexts[1]], progress / speed) / div;
+        color = mix(colors[int(indexes[1])], colors[int(nexts[1])], progress / speed) / div;
     }
     else if(color.b == 253.0 / 255.0){
-        color = mix(colors[indexes[2]], colors[nexts[2]], progress / speed) / div;
+        color = mix(colors[int(indexes[2])], colors[int(nexts[2])], progress / speed) / div;
     }
     else if(color.b == 252.0 / 255.0){
-        color = mix(colors[indexes[3]], colors[nexts[3]], progress / speed) / div;
+        color = mix(colors[int(indexes[3])], colors[int(nexts[3])], progress / speed) / div;
     }
     else if(color.b == 251.0 / 255.0){
-        color = mix(colors[indexes[4]], colors[nexts[4]], progress / speed) / div;
+        color = mix(colors[int(indexes[4])], colors[int(nexts[4])], progress / speed) / div;
     }
     else if(color.b == 250.0 / 255.0){
-        color = mix(colors[indexes[5]], colors[nexts[5]], progress / speed) / div;
+        color = mix(colors[int(indexes[5])], colors[int(nexts[5])], progress / speed) / div;
     }
 
     /*
     if(color.r == 1.0){
-        color = mix(colors[index], colors[next], progress / 60.0) / 1.2;
+        color = mix(colors[int(index)], colors[int(next)], progress / 60.0) / 1.2;
     }
     */
     gl_FragColor = color;
